Changed back to original
This should still be alarm_list_delete(alarmid), since we always search
from the head of the list anyway, so it's still O(1). I didn't realize
that in time so I didn't submit it :(
The alarmid parameter allows us to deregister other alarms not at the
head of the list.